Backend/requirements.txt
torch
transformers
scikit-learn
pysolr
nltk
google-generativeai
flask
flask_cors